Add files via upload
The Carrier Vehicle Travelling Salesman Problem is an extension of the classic Travelling Salesman Problem. In this method, ESG emissions is minimized simultaneously with the maximization of revenue and rating. This represents a much more difficult way of solving the use case of F1 calendar optimization, as the program must schedule the optimal circuits as the same time as it selects which 21 of the 30 circuits to include in the calendar. The complexity of this algorithm was beyond the scope of the course. It was not taken as the final solution to the use case as the constraints were not formulated correctly. Nevertheless, this code demonstrates a different - and arguably more suitable - method to solving the mandate.