Rename MGSC 662 Decision Analytics - Group 9 Final Project Code (Final Version Using MIP & TSP) (1).ipynb to F1 Calendar Optimization Using MIP and TSP M3ethod
To Solve the use case, a Mixed Integer algorithm was used in tandem with a Travelling Salesman algorithm. First, the MIP algorithm selected which of the 30 circuits maximized season revenue and rating. One those 2 criteria were maximized, the selected circuits were then scheduled in an order than minimized ESG emissions using a classic TSP algorithm. While this does present a suitable solution to the use case, the solution isn't the most optimal as the maximization of revenue and rating was prioritized over the minimization of ESG. Ideally, these two tasks should be executed simultaneously to yield an optimal solution